+++ R/Run_Mod.R
@@ -19,12 +19,14 @@
 #' @param nburn An [integer] specifying the number of burn ins before sampling.
 #' @param nmc An [integer] specifying number of samples in posterior chains
 #' @param niter An [integer] specifying number of iterations.
-#' @param strat
-#' A [character] string specifying the stratification variable,
+#' @param strat A [character] string specifying the stratification variable,
 #' entered in quotes.
+#' @param with_post A [logical] value specifying if raw jags post objects
+#' should be included in the output. Note: These objects can be large.
+#' @param include_subs A [logical] value specifying if posterior distributions
+#' should be included for all subjects. A value of [FALSE] will only include
+#' the predictive distribution.
 #' @return
-#' - A jags.post [list()] object or multiple jags.post [list()]
-#' if stratified. Returned as a [list()] of class [runjags::runjags-class]
 #' - A [base::data.frame()] titled `curve_params` that contains the posterior
 #' distribution will be exported with the following attributes:
 #'  - `iteration` = number of sampling iterations.
@@ -53,7 +55,9 @@ run_mod <- function(data,
                     nburn = 0,
                     nmc = 100,
                     niter = 100,
-                    strat = NA) {
+                    strat = NA,
+                    with_post = FALSE,
+                    include_subs = FALSE) {
   ## Conditionally creating a stratification list to loop through
   if (is.na(strat) == FALSE) {
     strat_list <- unique(data[[strat]])
@@ -114,13 +118,11 @@ run_mod <- function(data,
       summarise = FALSE
     )
     # Assigning the raw jags output to a list.
-    # This will include a raw output for the jags.post for each stratification.
+    # This will include a raw output for the jags.post for each stratification
+    # and will only be included if specified. 
     jags_post_final[[i]] <- jags_post
 
-    ## Cleaning the jags output -- much of this has to do with correctly
-    # classifying the [x,x] number
-    # included in the outputs
-    # ggs works with mcmc objects
+    # Unpacking and cleaning mcmc output.
     jags_unpack <- ggmcmc::ggs(jags_post[["mcmc"]])
 
     # Adding attributes
@@ -132,7 +134,7 @@ run_mod <- function(data,
     # then by the order they are estimated by the program.
     iso_dat <- data.frame(attributes(longdata)$antigens)
     iso_dat <- iso_dat |> dplyr::mutate(Subnum = as.numeric(row.names(iso_dat)))
-    ### Working with jags unpacked ggs outputs to clarify parameter and subject
+    # Working with jags unpacked ggs outputs to clarify parameter and subject
     jags_unpack <- jags_unpack |>
       dplyr::mutate(
         Subnum = sub(".*,", "", .data$Parameter),
@@ -145,14 +147,14 @@ run_mod <- function(data,
       )
     # Merging isodat in to ensure we are classifying antigen_iso
     jags_unpack <- dplyr::left_join(jags_unpack, iso_dat, by = "Subnum")
-    jags_unpack <- jags_unpack |>
-      dplyr::rename(c("Iso_type" = "attributes.longdata..antigens")) |>
-      dplyr::select(!c("Subnum"))
-    # Setting subset for the "new person" --setting it to the final sample
-    np <- as.character(longdata$nsubj)
+    ids <- data.frame(attr(longdata, "ids")) |>
+      mutate(Subject = as.character(dplyr::row_number()))
+    jags_unpack <- dplyr::left_join(jags_unpack, ids, by = "Subject")
     jags_final <- jags_unpack |>
-      dplyr::filter(.data$Subject == np)
-    ## Creating a label for the stratification, if there is one.
+      dplyr::select(!c("Subnum", "Subject")) |>
+      dplyr::rename(c("Iso_type" = "attributes.longdata..antigens",
+                      "Subject" = "attr.longdata...ids.."))
+    # Creating a label for the stratification, if there is one.
     # If not, will add in "None".
     jags_final$Stratification <- i
     ## Creating output
@@ -161,12 +163,26 @@ run_mod <- function(data,
   # Ensuring output does not have any NAs
   jags_out <- jags_out[complete.cases(jags_out), ]
   # Outputting the finalized jags output as a data frame with the
-  # jags output results for each stratification
-  # rbinded.
-  jags_out <- list(
-    "curve_params" = jags_out,
-    "jags.post" = jags_post_final,
-    "attributes" = mod_atts
-  )
+  # jags output results for each stratification rbinded.
+  # Logical argument to include posterior of all subjects or just the
+  # predictive distribution (new person).
+  if (include_subs == FALSE) {
+    jags_out <- jags_out |>
+      filter(.data$Subject == "newperson")
+  }
+
+  # Logical argument to keep the raw jags post or not.
+  if (with_post == TRUE) {
+    jags_out <- list(
+      "curve_params" = jags_out,
+      "jags.post" = jags_post_final,
+      "attributes" = mod_atts
+    )
+  } else { 
+    jags_out <- list(
+      "curve_params" = jags_out,
+      "attributes" = mod_atts
+    )
+  }
   jags_out
 }