+++ inst/examples/examples-plot_jags_effectivedx.R
@@ -1,5 +1,5 @@
 
-data <- serodynamics::nepal_sees_jags_post
+data <- serodynamics::nepal_sees_jags_output
 
 plot_jags_effect(data = data,
                  iso = "HlyE_IgA",