+++ inst/examples/examples-plot_jags_rhatdx.R
@@ -1,5 +1,5 @@
 
-data <- serodynamics::nepal_sees_jags_post
+data <- serodynamics::nepal_sees_jags_output
 
 plot_jags_Rhat(data = data,
                iso = "HlyE_IgA",