+++ inst/examples/examples-plot_jags_tracedx.R
@@ -1,5 +1,5 @@
 
-data <- serodynamics::nepal_sees_jags_post
+data <- serodynamics::nepal_sees_jags_output
 
 # Specifying isotype, parameter, and stratification for traceplot.
 plot_jags_trace(