+++ tests/testthat/test-plot_jags_rhatdx.R
@@ -4,7 +4,7 @@ test_that(
   code = {
     skip_if(getRversion() < "4.4.1") # 4.3.3 had issues
 
-    data <- serodynamics::nepal_sees_jags_post |>
+    data <- serodynamics::nepal_sees_jags_output |>
       suppressWarnings()
 
     # Testing for any errors: